+++ trantor/utils/crypto/sha256.cc
@@ -29,7 +29,7 @@
 #define SIG1(x) (ROTRIGHT(x, 17) ^ ROTRIGHT(x, 19) ^ ((x) >> 10))
 
 /**************************** VARIABLES *****************************/
-static const unsigned int k[64] = {
+static const uint32_t k[64] = {
     0x428a2f98, 0x71374491, 0xb5c0fbcf, 0xe9b5dba5, 0x3956c25b, 0x59f111f1,
     0x923f82a4, 0xab1c5ed5, 0xd807aa98, 0x12835b01, 0x243185be, 0x550c7dc3,
     0x72be5d74, 0x80deb1fe, 0x9bdc06a7, 0xc19bf174, 0xe49b69c1, 0xefbe4786,
@@ -43,9 +43,9 @@ static const unsigned int k[64] = {
     0x90befffa, 0xa4506ceb, 0xbef9a3f7, 0xc67178f2};
 
 /*********************** FUNCTION DEFINITIONS ***********************/
-void trantor_sha256_transform(SHA256_CTX *ctx, const BYTE data[])
+void trantor_sha256_transform(SHA256_CTX *ctx, const uint8_t data[])
 {
-    WORD a, b, c, d, e, f, g, h, i, j, t1, t2, m[64];
+    uint32_t a, b, c, d, e, f, g, h, i, j, t1, t2, m[64];
 
     for (i = 0, j = 0; i < 16; ++i, j += 4)
         m[i] = (data[j] << 24) | (data[j + 1] << 16) | (data[j + 2] << 8) |
@@ -100,9 +100,9 @@ void trantor_sha256_init(SHA256_CTX *ctx)
     ctx->state[7] = 0x5be0cd19;
 }
 
-void trantor_sha256_update(SHA256_CTX *ctx, const BYTE data[], size_t len)
+void trantor_sha256_update(SHA256_CTX *ctx, const uint8_t data[], size_t len)
 {
-    WORD i;
+    uint32_t i;
 
     for (i = 0; i < len; ++i)
     {
@@ -117,9 +117,9 @@ void trantor_sha256_update(SHA256_CTX *ctx, const BYTE data[], size_t len)
     }
 }
 
-void trantor_sha256_final(SHA256_CTX *ctx, BYTE hash[])
+void trantor_sha256_final(SHA256_CTX *ctx, uint8_t hash[])
 {
-    WORD i;
+    uint32_t i;
 
     i = ctx->datalen;
 
@@ -141,14 +141,14 @@ void trantor_sha256_final(SHA256_CTX *ctx, BYTE hash[])
 
     // Append to the padding the total message's length in bits and transform.
     ctx->bitlen += ctx->datalen * 8;
-    ctx->data[63] = ctx->bitlen;
-    ctx->data[62] = ctx->bitlen >> 8;
-    ctx->data[61] = ctx->bitlen >> 16;
-    ctx->data[60] = ctx->bitlen >> 24;
-    ctx->data[59] = ctx->bitlen >> 32;
-    ctx->data[58] = ctx->bitlen >> 40;
-    ctx->data[57] = ctx->bitlen >> 48;
-    ctx->data[56] = ctx->bitlen >> 56;
+    ctx->data[63] = (uint8_t)ctx->bitlen;
+    ctx->data[62] = (uint8_t)(ctx->bitlen >> 8);
+    ctx->data[61] = (uint8_t)(ctx->bitlen >> 16);
+    ctx->data[60] = (uint8_t)(ctx->bitlen >> 24);
+    ctx->data[59] = (uint8_t)(ctx->bitlen >> 32);
+    ctx->data[58] = (uint8_t)(ctx->bitlen >> 40);
+    ctx->data[57] = (uint8_t)(ctx->bitlen >> 48);
+    ctx->data[56] = (uint8_t)(ctx->bitlen >> 56);
     trantor_sha256_transform(ctx, ctx->data);
 
     // Since this implementation uses little endian byte ordering and SHA uses
